subiendo actualizaciones

+++ Entities/Cargo.cs
@@ -5,7 +5,7 @@ namespace WebApiKalum.Entities
         public string CargoId { get; set; }
         public string Descripcion { get; set; }
         public string Prefijo { get; set; }
-        public double  Monto { get; set; } //double utilizado para declarar decimal
+        public decimal Monto { get; set; } //double utilizado para declarar decimal
         public bool GeneraMora { get; set; }//bool utilizado para declarar bit
         public int PorcentajeMora { get; set; } //int utilizado para declarar enteros(int)
